Cybersec Flask App/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
openai>=1.40.0
requests>=2.32.3